+++ lib/glimmer/launcher.rb
@@ -19,7 +19,10 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'bundler/setup' if ARGV.include?('--bundler') && File.exist?(File.expand_path('./Gemfile'))
+if ARGV.include?('--bundler') && File.exist?(File.expand_path('./Gemfile'))
+  require 'bundler'
+  Bundler.setup(:default)
+end
 require 'fileutils'
 require 'os'
 
@@ -39,8 +42,8 @@ module Glimmer
       Optionally, extra Glimmer options, JRuby options, and/or environment variables may be passed in.
     
       Glimmer options:
-      - "--bundler"         : Activates Glimmer gems with `bundler/setup` to ensure they match Gemfile (passes -G to jruby on app run)
-      - "--quiet"           : Does not announce file path of Glimmer application being launched
+      - "--bundler=GROUP"   : Activates gems in Bundler default group in Gemfile
+      - "--quiet=BOOLEAN"   : Does not announce file path of Glimmer application being launched
       - "--debug"           : Displays extra debugging information, passes "--debug" to JRuby, and enables debug logging
       - "--log-level=VALUE" : Sets Glimmer's Ruby logger level ("ERROR" / "WARN" / "INFO" / "DEBUG"; default is none)
     
@@ -54,7 +57,8 @@ module Glimmer
     GLIMMER_LIB_GEM = 'glimmer-dsl-swt'
     GLIMMER_OPTIONS = %w[--log-level --quiet --bundler]
     GLIMMER_OPTION_ENV_VAR_MAPPING = {
-      '--log-level' => 'GLIMMER_LOGGER_LEVEL'
+      '--log-level' => 'GLIMMER_LOGGER_LEVEL'   ,
+      '--bundler'   => 'GLIMMER_BUNDLER_SETUP'  ,
     }
     REGEX_RAKE_TASK_WITH_ARGS = /^([^\[]+)\[?([^\]]*)\]?$/
 
@@ -105,7 +109,6 @@ module Glimmer
 
       def launch(application, jruby_options: [], env_vars: {}, glimmer_options: {})
         jruby_options_string = jruby_options.join(' ') + ' ' if jruby_options.any?
-        bundler_require = '-G' if glimmer_options['--bundler']
         env_vars = env_vars.merge(glimmer_option_env_vars(glimmer_options))
         env_vars_string = env_vars.map do |k,v| 
           if OS.windows? && ENV['PROMPT'] # detect command prompt (or powershell)
@@ -137,7 +140,7 @@ module Glimmer
           @@mutex.synchronize do
             puts "Launching Glimmer Application: #{application}" if jruby_options_string.to_s.include?('--debug') || glimmer_options['--quiet'].to_s.downcase != 'true'
           end
-          command = "#{env_vars_string} jruby #{bundler_require} #{jruby_options_string}#{jruby_os_specific_options} #{devmode_require}-r #{the_glimmer_lib} -S #{application}"
+          command = "#{env_vars_string} jruby #{jruby_options_string}#{jruby_os_specific_options} #{devmode_require}-r #{the_glimmer_lib} -S #{application}"
           if !env_vars_string.empty? && OS.windows?
             command = "bash -c \"#{command}\"" if ENV['SHELL'] # do in Windows Git Bash only
             command = "cmd /C \"#{command}\"" if ENV['PROMPT'] # do in Windows Command Prompt only (or Powershell)